Try to fix random errors with Symfony by not using anonymous classes
